add script to measure gas for stable tradiing strategy

+++ contracts/strategies/lib/Stable.sol
@@ -14,6 +14,9 @@
 
 pragma solidity ^0.7.1;
 
+import "hardhat/console.sol";
+
+
 import "../../math/FixedPoint.sol";
 
 // This is a contract to emulate file-level functions. Convert to a library
@@ -23,14 +26,14 @@ import "../../math/FixedPoint.sol";
 // solhint-disable var-name-mixedcase
 
 contract Stable {
-    uint256 internal constant PRECISION = 100000000000000;
+    int256 internal constant PRECISION = 100000000000000;
 
     struct Data {
-        uint256 amp;
-        uint256 invariant;
-        uint256 sum;
-        uint256 nn;
-        uint256 prod;
+        int256 amp;
+        int256 invariant;
+        int256 sum;
+        int256 nn;
+        int256 prod;
     }
 
     function _approximateAmount(Data memory data, uint128 approxAmount)
@@ -39,21 +42,21 @@ contract Stable {
         returns (uint128)
     {
         uint128 newApproxAmount;
-        uint256 c1 = data.amp *
+        int256 c1 = data.amp *
             data.sum +
             ((FixedPoint.ONE / data.nn) - data.amp) *
             data.invariant;
-        uint256 c2 = (data.invariant * data.invariant * data.invariant) /
+        int256 c2 = (data.invariant * data.invariant * data.invariant) /
             (data.nn * data.nn * data.prod);
-        for (uint256 i = 0; i < 255; i++) {
-            uint256 f1 = data.amp *
+        for (int256 i = 0; i < 255; i++) {
+            int256 f1 = data.amp *
                 approxAmount *
                 approxAmount +
                 c1 *
                 approxAmount -
                 c2 *
                 FixedPoint.ONE;
-            uint256 f2 = c1 + 2 * data.amp * approxAmount;
+            int256 f2 = c1 + 2 * data.amp * approxAmount;
             newApproxAmount = uint128(approxAmount - (f1 / f2));
             if (newApproxAmount > approxAmount) {
                 if ((newApproxAmount - approxAmount) <= PRECISION) {
@@ -68,17 +71,17 @@ contract Stable {
     }
 
     function _getDataOutGivenIn(
-        uint256 amp,
+        int256 amp,
         uint128[] memory balances,
         uint256 tokenIndexIn,
         uint256 tokenIndexOut,
         uint128 tokenAmountIn
     ) private pure returns (Data memory) {
-        uint256 invariant = _invariant(amp, balances);
-        uint256 sum = 0;
-        uint256 prod = FixedPoint.ONE;
+        int256 invariant = _invariant(amp, balances);
+        int256 sum = 0;
+        int256 prod = FixedPoint.ONE;
         uint256 n = balances.length;
-        uint256 nn = 1;
+        int256 nn = 1;
         for (uint256 i = 0; i < n; i++) {
             if (i != tokenIndexOut) {
                 if (i == tokenIndexIn) {
@@ -91,7 +94,7 @@ contract Stable {
                     prod = (prod * balances[i]) / FixedPoint.ONE;
                 }
             }
-            nn = nn * n;
+            nn = nn * int256(n);
         }
         return
             Data({
@@ -104,17 +107,17 @@ contract Stable {
     }
 
     function _getDataInGivenOut(
-        uint256 amp,
+        int256 amp,
         uint128[] memory balances,
         uint256 tokenIndexIn,
         uint256 tokenIndexOut,
         uint128 tokenAmountOut
     ) private pure returns (Data memory) {
-        uint256 invariant = _invariant(amp, balances);
-        uint256 sum = 0;
-        uint256 prod = FixedPoint.ONE;
+        int256 invariant = _invariant(amp, balances);
+        int256 sum = 0;
+        int256 prod = FixedPoint.ONE;
         uint256 n = balances.length;
-        uint256 nn = 1;
+        int256 nn = 1;
         for (uint256 i = 0; i < n; i++) {
             if (i != tokenIndexIn) {
                 if (i == tokenIndexOut) {
@@ -127,7 +130,7 @@ contract Stable {
                     prod = (prod * balances[i]) / FixedPoint.ONE;
                 }
             }
-            nn = nn * n;
+            nn = nn * int256(n);
         }
         return
             Data({
@@ -140,7 +143,7 @@ contract Stable {
     }
 
     function _outGivenIn(
-        uint256 amp,
+        int256 amp,
         uint128[] memory balances,
         uint256 tokenIndexIn,
         uint256 tokenIndexOut,
@@ -160,7 +163,7 @@ contract Stable {
     }
 
     function _inGivenOut(
-        uint256 amp,
+        int256 amp,
         uint128[] memory balances,
         uint256 tokenIndexIn,
         uint256 tokenIndexOut,
@@ -179,31 +182,31 @@ contract Stable {
             balances[tokenIndexIn];
     }
 
-    function _invariant(uint256 amp, uint128[] memory balances)
+    function _invariant(int256 amp, uint128[] memory balances)
         internal
         pure
-        returns (uint256)
+        returns (int256)
     {
-        uint256 sum = 0;
-        uint256 prod = FixedPoint.ONE;
+        int256 sum = 0;
+        int256 prod = FixedPoint.ONE;
         uint256 n = balances.length;
-        uint256 nn = 1;
+        int256 nn = 1;
         for (uint256 i = 0; i < n; i++) {
             sum = sum + balances[i];
             prod = (prod * balances[i]) / FixedPoint.ONE;
-            nn = nn * n;
+            nn = nn * int256(n);
         }
-        uint256 invariant = sum;
-        uint256 newInvariant;
-        uint256 c2 = amp - FixedPoint.ONE / nn;
-        uint256 c1 = (nn * nn * prod);
-        for (uint256 i = 0; i < 255; i++) {
-            uint256 f1 = (c2 *
+        int256 invariant = sum;
+        int256 newInvariant;
+        int256 c2 = amp - FixedPoint.ONE / nn;
+        int256 c1 = (nn * nn * prod);
+        for (int256 i = 0; i < 255; i++) {
+            int256 f1 = (c2 *
                 invariant +
                 (((invariant * invariant) / c1) * invariant) -
                 amp *
                 sum) / FixedPoint.ONE;
-            uint256 f2 = (c2 *
+            int256 f2 = (c2 *
                 FixedPoint.ONE +
                 3 *
                 ((invariant * FixedPoint.ONE) / c1) *
